+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Each change should be on one line, as GitHub markdown in the release section can
 
 - `metalarchives` plugin: fetches lyrics from metal-archives.com (Encyclopaedia Metallum).
 - `songlyrics` plugin: fetches lyrics from songlyrics.com.
+- `utatime` plugin: fetches lyrics from utatime.com (formerly lyrical-nonsense.com).
 
 ## [1.0.0] - 2026-08-04